Add files via upload
according to the llm ,we write these docs to describe the flaws of  llms ,and we are making an agent to improve their actions in answering ai rebuttals or reviews. There are 2 questions llms face ,the first is hallucination, llms tend to give answers with no evidences in papers; the second is llms are willing to promise to do things even it's hard to accomplish. For example, maybe an experiment is perfect in a accurate area, you can't do better than it. So it's no use promising like we will do a new experiment to transcend the former experiment.